Fix for Issue 44: https://code.google.com/p/python-xbee/issues/detail?id=44
Thanks to Martin Erablitek for the report. The software mask of the DIO mask
value in the sample data header for XBee ZB (ZigBee) devices was incorrect;
this would lead to incorrect parsing of sample data. As a result, DIO lines
12 and 4 could never be parsed as enabled. In addition, if these were the
only enabled lines, any following ADC data would be garbage.

The fix was to update the software mask for the DIO mask value from 0x0E7F
to the proper value 0x1CFF. In addition, test cases were updated to
demonstrate correct parsing, and the changed test cases were manually
verified against the specification.